create api for cart and order

+++ src/main/java/com/langgomsport/langgomsport/repository/VariantRepository.java
@@ -3,6 +3,6 @@ package com.langgomsport.langgomsport.repository;
 import com.langgomsport.langgomsport.models.Variant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VariantRepository extends JpaRepository<Variant, Long> {
+public interface VariantRepository extends JpaRepository<Variant, Integer> {
 
 }
